Revert "split contact.page and base-form.page"
This reverts commit 68d3efe5ea264bb3474bb35c82b50ea3c6a2450e.
